Remove unnecessary delegates
